+++ src/Models/MonitoredScheduledTask.php
@@ -82,6 +82,10 @@ class MonitoredScheduledTask extends Model
 
     public function markAsFinished(ScheduledTaskFinished $event): self
     {
+        if ($this->eventConcernsBackgroundTaskThatCompletedInForeground($event)) {
+            return $this;
+        }
+
         if ($event->task->exitCode !== 0 && ! is_null($event->task->exitCode)) {
             return $this->markAsFailed($event);
         }
@@ -89,9 +93,9 @@ class MonitoredScheduledTask extends Model
         $logItem = $this->createLogItem(MonitoredScheduledTaskLogItem::TYPE_FINISHED);
 
         $logItem->updateMeta([
-            'runtime' => $event->runtime,
+            'runtime' => $event->task->runInBackground ? 0 : $event->runtime,
             'exit_code' => $event->task->exitCode,
-            'memory' => memory_get_usage(true),
+            'memory' => $event->task->runInBackground ? 0 : memory_get_usage(true),
         ]);
 
         $this->update(['last_finished_at' => now()]);
@@ -99,6 +103,15 @@ class MonitoredScheduledTask extends Model
         $this->pingOhDear($logItem);
 
         return $this;
+    }
+
+    public function eventConcernsBackgroundTaskThatCompletedInForeground(ScheduledTaskFinished $event): bool
+    {
+        if (! $event->task->runInBackground) {
+            return false;
+        }
+
+        return $event->task->exitCode === null;
     }
 
     /**
@@ -149,7 +162,7 @@ class MonitoredScheduledTask extends Model
         if (! in_array($logItem->type, [
             MonitoredScheduledTaskLogItem::TYPE_FAILED,
             MonitoredScheduledTaskLogItem::TYPE_FINISHED,
-        ])) {
+        ], true)) {
             return $this;
         }