+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to `laravel-schedule-monitor` will be documented in this file
 
+## 2.2.0 - 2021-01-15
+
+- throw an exception if pinging Oh Dear has failed [#37](https://github.com/spatie/laravel-schedule-monitor/pull/37)
+- pass 0 instead of null parameters to Oh dear for Background tasks [#37](https://github.com/spatie/laravel-schedule-monitor/pull/37)
+
+## 2.1.0 - 2020-12-04
+
+- add support for PHP 8
+
+## 2.0.2 - 2020-10-14
+
+- drop support for Laravel 7
+- fix command description
+
+## 2.0.1 - 2020-10-06
+
+- report right exit code for scheduled tasks in background
+
+## 2.0.0 - 2020-09-29
+
+- add support for timezones
+
+## 1.0.4 - 2020-09-08
+
+- add support for Laravel 8
+
 ## 1.0.3 - 2020-07-14
 
 - fix link config file